block: add support for generic mount options
Not all mount options are to be passed directly to the filesystem driver,
some of them are generic VFS flags that must be passed as bitfield to the
mount() syscall.

Add a new option parsing implementation that properly divides the user
sepcified mount options into flags and fs specific arguments.